% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_body.R
\name{auth_body}
\alias{auth_body}
\title{Create the body for the auth part of the dashboard
This will differ diffrent dependent upon the status of the app}
\usage{
auth_body(input, output, session, status)
}
\arguments{
\item{status}{What is the status of the app and as such which of the sidebars should be shown}
}
\description{
Create the body for the auth part of the dashboard
This will differ diffrent dependent upon the status of the app
}
