% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_check.R
\name{auth_check}
\alias{auth_check}
\title{Check the username and password given in input$user and input$password respectivly in the current namespace
If the credentuals are correct return the user information}
\usage{
auth_check(input, output, session, auth)
}
\description{
Check the username and password given in input$user and input$password respectivly in the current namespace
If the credentuals are correct return the user information
}
