% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_server.R
\name{auth_server}
\alias{auth_server}
\title{Manage the logon, and the changes of the UI based upon}
\usage{
auth_server(server, config_path)
}
\arguments{
\item{server}{Shiny Server function to run when a user is logged on, must take aut as argument}

\item{auth_config}{File path to auth_config file}
}
\description{
Manage the logon, and the changes of the UI based upon
}
