% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_sidebar.R
\name{authSidebarMenu}
\alias{authSidebarMenu}
\title{Function that creates a sidebar with the id authMenuItems and shinyauth tabs automaticly
This should be used like shinydashboard::sidebarMenu exept the id cannot be changed}
\usage{
authSidebarMenu(auth, ...)
}
\description{
Function that creates a sidebar with the id authMenuItems and shinyauth tabs automaticly
This should be used like shinydashboard::sidebarMenu exept the id cannot be changed
}
