% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_tab.R
\name{change_user_admin_status}
\alias{change_user_admin_status}
\title{Function that allos for the awaring and revoking of admin rights}
\usage{
change_user_admin_status(input, output, session, auth, dt_user)
}
\description{
Function that allos for the awaring and revoking of admin rights
}
