% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_tab.R
\name{change_user_details}
\alias{change_user_details}
\title{Change users Deitals must be called from inside a name space}
\usage{
change_user_details(input, output, session, auth, dt_settings_user)
}
\description{
Change users Deitals must be called from inside a name space
}
