% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_auth.R
\name{check_user_id}
\alias{check_user_id}
\title{Given a user id return TRUE if they exist in the db and FALSE if not}
\usage{
check_user_id(auth, user_id)
}
\description{
Given a user id return TRUE if they exist in the db and FALSE if not
}
