% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_auth_tables.R
\name{create_auth_tables}
\alias{create_auth_tables}
\title{Function to create tables for shinyauth, this funciton will also chec that the
auth.config.yaml file contanis the requite information}
\usage{
create_auth_tables(auth_config_path)
}
\description{
Function to create tables for shinyauth, this funciton will also chec that the
auth.config.yaml file contanis the requite information
}
