% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user.R
\name{create_user_modal}
\alias{create_user_modal}
\title{User creation modal}
\usage{
create_user_modal(input, output, session, time_stamp)
}
\arguments{
\item{time_stamp}{Creates a uniqe box each time the box is opends}
}
\description{
User creation modal
}
