% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_password.R
\name{generate_random_password}
\alias{generate_random_password}
\title{Generate a random password using sodium}
\usage{
generate_random_password(n = 1, lenght = 11)
}
\description{
Generate a random password using sodium
}
