% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user.R
\name{get_curret_user_ids}
\alias{get_curret_user_ids}
\title{Query the db for the table of all of the user information}
\usage{
get_curret_user_ids(auth)
}
\description{
Query the db for the table of all of the user information
}
