% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_tab.R
\name{get_users_table}
\alias{get_users_table}
\title{Query the db for the table of all of the user information}
\usage{
get_users_table(input, output, session, auth)
}
\description{
Query the db for the table of all of the user information
}
