% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{getToken}
\alias{getToken}
\title{Performs a post request to retreive a user's token.}
\usage{
getToken(auth_code, redirect_uri, key, token_url)
}
\arguments{
\item{auth_code}{User's previously retreived authorization code.}

\item{redirect_uri}{your apps redirect URL (aka where you app lives on the internet or localhost)}

\item{key}{Your apps api key.}

\item{token_url}{the token retreival address for your api.}
}
\value{
A string formatted to be plopped into a \code(httr::POST) request.
}
\description{
Performs a post request to retreive a user's token.
}
\examples{
getToken('myrecorder')
}
