% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_auth_object.R
\name{make_auth_object}
\alias{make_auth_object}
\title{Make the auth object from the auth_config file and the connection}
\usage{
make_auth_object(config_path)
}
\arguments{
\item{config_path}{The path to the config file used by Shiny Basic Auth}
}
\description{
Make the auth object from the auth_config file and the connection
}
