% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBodyString.R
\name{makeBodyString}
\alias{makeBodyString}
\title{Assembles a body for a post request to retreive the user's token.}
\usage{
makeBodyString(key, redirect_uri, auth_code)
}
\arguments{
\item{key}{your personal api key}

\item{redirect_uri}{your apps redirect URL (aka where you app lives on the internet or localhost)}

\item{auth_code}{Code that was retreived from an earlier authorization request client-side.}
}
\value{
A string formatted to be plopped into a \code(httr::POST) request.
}
\description{
Assembles a body for a post request to retreive the user's token.
}
