% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_modal.R
\name{message_modal}
\alias{message_modal}
\title{Show a message to the user, without custom javascript}
\usage{
message_modal(input, output, session, title, message)
}
\description{
Show a message to the user, without custom javascript
}
