% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_password.R
\name{password_change_manager}
\alias{password_change_manager}
\title{This function creates and manages the life cycle of a uniqe password_change_modal that
allos the user to change there password.  Or an andmin to change the user passwords for
any user.}
\usage{
password_change_manager(input, output, session, auth, admin = FALSE,
  user_id = NULL, old_password = NULL, message = NULL)
}
\arguments{
\item{admin}{Is true if this is an admin changin the datils for another user}

\item{message}{Optional message to show in the password modal}
}
\description{
This function creates and manages the life cycle of a uniqe password_change_modal that
allos the user to change there password.  Or an andmin to change the user passwords for
any user.
}
