% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_password.R
\name{password_change_modal}
\alias{password_change_modal}
\title{Password Reset/Change modal ui box
The UI shown depends on if this is for the logged inuser or an admin
changeing the passowrd of anothe user}
\usage{
password_change_modal(input, output, session, admin, user_id, time_stamp,
  message = NULL)
}
\arguments{
\item{time_stamp}{Creates a uniqe box each time the box is opends}
}
\description{
Password Reset/Change modal ui box
The UI shown depends on if this is for the logged inuser or an admin
changeing the passowrd of anothe user
}
