% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_tab.R
\name{render_admin_page}
\alias{render_admin_page}
\title{Function creating the ui for a admin page}
\usage{
render_admin_page(input, output, session)
}
\description{
Function creating the ui for a admin page
}
