% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{saBody}
\alias{saBody}
\title{Deffine the ui for the body of the shiny app using shinyauth}
\usage{
saBody(...)
}
\arguments{
\item{...}{Passed to shinydashboard::dashboardBody}
}
\description{
Deffine the ui for the body of the shiny app using shinyauth
}
