% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_functions.R
\name{saSidebar}
\alias{saSidebar}
\title{Deffine the ui for the sidebar of the shiny app using shinyauth}
\usage{
saSidebar(...)
}
\arguments{
\item{...}{Passed to shinydashboard::dashboardSidebar}
}
\description{
Deffine the ui for the sidebar of the shiny app using shinyauth
}
