% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_password.R
\name{save_new_password}
\alias{save_new_password}
\title{Save a new password to the db}
\usage{
save_new_password(session, auth, user_id, user_changed, new_password)
}
\arguments{
\item{user_changed}{Has the password been changed by the user, or reset by admin}
}
\description{
Save a new password to the db
}
