% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_tab.R
\name{save_user_details}
\alias{save_user_details}
\title{Save the current user datailes to the db}
\usage{
save_user_details(input, output, session, auth, permissions, dt_user = NULL)
}
\arguments{
\item{dt_user}{data.table of user for whom the settings page is opened, defult so acitve user}
}
\description{
Save the current user datailes to the db
}
