% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_tab.R
\name{settings_tab}
\alias{settings_tab}
\title{Create the settings tab
This function needs to be able to trigger the fech of a new dt_user}
\usage{
settings_tab(input, output, session, auth, permissions = "user")
}
\arguments{
\item{dt_user}{Data table identifying the current user}
}
\description{
Create the settings tab
This function needs to be able to trigger the fech of a new dt_user
}
