% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyAuth_Keycloak.R
\name{ShinyAuth_Keycloak}
\alias{ShinyAuth_Keycloak}
\title{Authentication Class for Keycloak authentication}
\description{
Authentication Class for Keycloak authentication

Authentication Class for Keycloak authentication
}
\details{
Provides methods for protecting a Shiny app against Auth0 as identity provider.
}
\section{Super class}{
\code{\link[ShinyAuth:Abstract_ShinyAuth]{ShinyAuth::Abstract_ShinyAuth}} -> \code{ShinyAuth_Keycloak}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{url_auth_template}}{URL template for auth entrypoint.}

\item{\code{url_token_template}}{URL template for token entrypoint.}

\item{\code{url_userinfo_template}}{URL template for userinfo entrypoint.}

\item{\code{url_logout_template}}{URL template for logout entrypoint.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ShinyAuth_Keycloak$new()}}
\item \href{#method-clone}{\code{ShinyAuth_Keycloak$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="app">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-app}{\code{ShinyAuth::Abstract_ShinyAuth$app()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="generate_state">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-generate_state}{\code{ShinyAuth::Abstract_ShinyAuth$generate_state()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="get_access_token">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-get_access_token}{\code{ShinyAuth::Abstract_ShinyAuth$get_access_token()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="get_userinfo">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-get_userinfo}{\code{ShinyAuth::Abstract_ShinyAuth$get_userinfo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="has_auth_code">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-has_auth_code}{\code{ShinyAuth::Abstract_ShinyAuth$has_auth_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="scripts">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-scripts}{\code{ShinyAuth::Abstract_ShinyAuth$scripts()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="server">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-server}{\code{ShinyAuth::Abstract_ShinyAuth$server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="ui">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-ui}{\code{ShinyAuth::Abstract_ShinyAuth$ui()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="url_auth">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-url_auth}{\code{ShinyAuth::Abstract_ShinyAuth$url_auth()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="url_logout">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-url_logout}{\code{ShinyAuth::Abstract_ShinyAuth$url_logout()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="url_token">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-url_token}{\code{ShinyAuth::Abstract_ShinyAuth$url_token()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ShinyAuth" data-topic="Abstract_ShinyAuth" data-id="url_userinfo">}\href{../../ShinyAuth/html/Abstract_ShinyAuth.html#method-url_userinfo}{\code{ShinyAuth::Abstract_ShinyAuth$url_userinfo()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize Keycloak authentication for shiny app
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyAuth_Keycloak$new(
  client_id,
  client_secret,
  app_url,
  auth_domain,
  realm,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_id}}{Client ID from IdP}

\item{\code{client_secret}}{Client Secret from IdP}

\item{\code{app_url}}{This application entrypoint url}

\item{\code{auth_domain}}{The domain at which to reach authentication server}

\item{\code{realm}}{Keycloak user realm}

\item{\code{...}}{Other parameters that will get stored in Service Provider settings. Can be used in url templates and in provider classes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new authentication object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyAuth_Keycloak$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
