% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_auth.R
\name{test_auth}
\alias{test_auth}
\title{Make an auth object like the one that would be present after a sucessful login,
To this end the funciton can  create an auth object of either no-login or the login of a given
user.  However, the defult behavoir of the function is create auth for the 'first_user' given
in the config file}
\usage{
test_auth(config_path, user_id = NULL, no_user = FALSE)
}
\arguments{
\item{config_path}{Path to config.yaml used by shinyauth}

\item{user_id}{The user id to create a loged in auth object for, if null defults to first_user}

\item{no_user}{Will override user_id and create an auth object for no user loged in}
}
\description{
Make an auth object like the one that would be present after a sucessful login,
To this end the funciton can  create an auth object of either no-login or the login of a given
user.  However, the defult behavoir of the function is create auth for the 'first_user' given
in the config file
}
