% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user.R
\name{user_creation_manager}
\alias{user_creation_manager}
\title{This funciton allows an admin to create a new user of the app}
\usage{
user_creation_manager(input, output, session, auth)
}
\description{
This funciton allows an admin to create a new user of the app
}
