% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_tab.R
\name{users_table_creation}
\alias{users_table_creation}
\title{Create table of users with buttons to change password and user details}
\usage{
users_table_creation(input, output, session, auth, dt_users)
}
\description{
Create table of users with buttons to change password and user details
}
