% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyauth.R
\name{shinyauth}
\alias{shinyauth}
\title{Server side component. You supply this with your api credentials.}
\usage{
shinyauth(input, output, session, api_url, api_key, scope)
}
\arguments{
\item{input}{you can ignore this as it is taken care of by shiny}

\item{output}{you can ignore this as it is taken care of by shiny}

\item{session}{you can ignore this as it is taken care of by shiny}

\item{api_url}{destination of the api you're querying from. For instance the fitbit api is 'https://www.fitbit.com/oauth2/authorize'}

\item{api_key}{your apps personal api key.}

\item{scope}{a vector of what you're requesting access to in the API. See the given api docs for examples.}
}
\description{
Server side component. You supply this with your api credentials.
}
\examples{
 drawChart <- callModule(shinydrawr,
    "outbreak_stats",
    random_data,
    draw_start = 15,
    x_key = "time",
    y_key = "metric")
}
